requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.1
django-jazzmin==2.6.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.4.4
